Normalise data tables

Harmonise and integrate data tables into standardised format

normTable(
  input = NULL,
  pattern = NULL,
  ...,
  outType = "rds",
  update = FALSE,
  verbose = FALSE
)

Arguments

input

[character(1)]
path of the file to normalise. If this is left empty, all files at stage two as subset by pattern are chosen.

pattern

[character(1)]
an optional regular expression. Only dataset names which match the regular expression will be processed.

...

[character(1)]
argument-value combination to filter the new tables by.

outType

[logical(1)]
the output file-type, currently implemented options are either *.csv (more exchangeable for a workflow based on several programs) or *.rds (smaller and less error-prone data-format but can only be read by R efficiently).

update

[logical(1)]
whether or not the physical files should be updated (TRUE) or the function should merely return the new object (FALSE, default). This is helpful to check whether the metadata specification and the provided file(s) (translation and ID tables) are properly specified.

verbose

[logical(1)]
be verbose about translating terms (default FALSE). Furthermore, you can use suppressMessages to make this function completely silent.

Value

This function harmonises and integrates so far unprocessed data tables at stage two into stage three of the areal database. It produces for each main polygon (e.g. nation) in the registered data tables a file that includes all thematic areal data.

Details

To normalise data tables, this function proceeds as follows:

Read in input and extract initial metadata from the file name.

Employ the function tabshiftr::reorganise to reshape input according to the respective schema description.

The territorial names are matched with the gazetteer to harmonise new territorial names (at this step, the function might ask the user to edit the file 'matching.csv' to align new names with already harmonised names).

Harmonise territorial unit names.

If update = TRUE, store the processed data table at stage three.

See also

Examples

if(dev.interactive()){
  # build the example database
  makeExampleDB(until = "normGeometry", path = tempdir())

  # normalise all available data tables ...
  normTable(update = TRUE)

  # ... and check the result
  output <- readRDS(paste0(tempdir(), "/adb_tables/stage3/Estonia.rds"))
}