Register a new dataseries

This function registers a new dataseries of both, geometries or areal data into the geospatial database. This contains the name and relevant meta-data of a dataseries to enable provenance tracking and reproducability.

regDataseries(
  name = NULL,
  description = NULL,
  homepage = NULL,
  version = NULL,
  licence_link = NULL,
  reference = NULL,
  notes = NULL,
  overwrite = FALSE
)

Arguments

name

[character(1)]
the dataseries abbreviation or name.

description

[character(1)]
the "long name" or "brief description" of the dataseries.

homepage

[character(1)]
the homepage of the data provider where the dataseries or additional information can be found.

version

[character(1)]
the version number or date when meta data of the dataseries were recorded.

licence_link

[character(1)]
link to the licence or the webpage from which the licence was copied.

reference

[bibentry(.)]
in case the dataseries comes with a reference, provide this here as bibentry object.

notes

[character(1)]
optional notes.

overwrite

[logical(1)]
whether or not the dataseries to register shall overwrite a potentially already existing older version.

Value

Returns a tibble of the new entry that is appended to 'inv_dataseries.csv'.

See also

Examples

if(dev.interactive()){
  # start the example database
  adb_exampleDB(until = "match_gazetteer", path = tempdir())

  regDataseries(name = "gadm",
                description = "Database of Global Administrative Areas",
                version = "3.6",
                homepage = "https://gadm.org/index.html",
                licence_link = "https://gadm.org/license.html")
}