Register a new areal data table

This function registers a new areal data table into the geospatial database.

regTable(nation = NULL, subset = NULL, dSeries = NULL,
  gSeries = NULL, level = NULL, begin = NULL, end = NULL,
  schema = NULL, archive = NULL, archiveLink = NULL,
  nextUpdate = NULL, updateFrequency = NULL, metadataLink = NULL,
  metadataPath = NULL, notes = NULL, update = FALSE,
  overwrite = FALSE)

Arguments

nation
[character(1)]
the nation for which the areal data are valid.

subset
[character(1)]
optional argument to specify which subset the file contains. This could be a subset of territorial units (e.g. only one municipality) or of a target variable.

dSeries
[character(1)]
the dataseries of the areal data.

gSeries
[character(1)]
optinally, the dataseries of the geometries, if the geometry dataseries deviates from the dataseries of the areal data.

level
[integerish(1)]
the administrative level at which the boundaries are recorded.

begin
[integerish(1)]
the date from which on the boundaries are valid.

end
[integerish(1)]
the date until which the boundaries are valid.

schema
[list(1)]
the schema description of the table to read in (must have been placed in the global environment before calling it here.

archive
[character(1)]
the original file from which the boundaries emerge.

archiveLink
[character(1)] download-link of the archive.

nextUpdate
[character(1)]
when does the geometry dataset gets updated the next time (format resricted to: YYYY-MM-DD).

updateFrequency
[character(1)]
value describing the frequency in which the dataset is updated, according to the ISO 19115 Codelist, MD_MaintenanceFrequencyCode. Possible values are: 'continual', 'daily', 'weekly', 'fortnightly', 'quarterly', 'biannually', 'annually', 'asNeeded', 'irregular', 'notPlanned', 'unknown', 'periodic', 'semimonthly', 'biennially'.

metadataLink
[character(1)]
if there is already metadata existing: link to the metadataset .

metadataPath
[character(1)]
if a existing metadataset was downloaded along the data: the path where it is stored locally.

notes
[character(1)]
optional notes.

update
[logical(1)]
whether or not the file 'inv_tables.csv' should be updated.

overwrite
[logical(1)]
whether or not the geometry to register shall overwrite a potentially already existing older version.

Value

Returns the entry that is appended to 'inv_tables.csv' in case update = TRUE.

Details

When processing areal data tables, carry out the following steps:

Determine the nation, administrative level, a subset (if applicable) and the dataseries of the areal data and of the geometry, and provide them as arguments to this function.

Provide a begin and end date for the areal data.

Run the function.

(Re)Save the table with the following properties:

Encoding: UTF-8

File name: What is provided as message by this function

make sure that the file is not modified or reshaped. This will happen during data normalisation via the schema description, which expects the original table.

Confirm that you have saved the file.

Every areal data dataseries (dSeries) may come as a slight permutation of a particular table arrangement. The function normTable expects internally a schema description (a list that describes the position of the data components) for each data table, which is saved as paste0("meta_", dSeries, TAB_NUMBER). A template thereof, and documentation on how to set them up, comes as the object tabshiftr::schema_default with arealDB.

Examples

if (FALSE) {

regTable(nation = "United States of America",
         subset = "soy",
         dSeries = "usda", gSeries = "gadm",
         level = 3,
         begin = 1990, end = 2017,
         archive = "soybean_us_county_1990_2017.csv",
         update = TRUE)
}